python/biomed_workflow/requirements.txt
databricks-sdk>=0.19.1
mlflow>=2.11
databricks-genai-inference>=0.2.1
databricks_vectorsearch==0.39
overrides>=7.7.0
nest-asyncio>=1.5.6
pkginfo>=1.10.0
defusedxml==0.7.1
delta-spark==3.2.0
beautifulsoup4==4.11.1
boto3==1.24.28
unstructured==0.14.9
html2text==2024.2.26